Added support for asset instancing.
As per #15, added support for the creation of instances of assets.
This provides the ability to create copies of existing assets that can be modified independantly (e.g. get a copy of a material so that changes to its properties don't affect other users of the material).
Instances use special instance urns and are disposed with the original asset.