Added support for reloading changes to assets.
This works only for exploded (directory) modules.

ModuleAssetProducer now establishes a watchService for each module, to monitor for changes to files relating to the assets it produces.  To ensure correct behavior, this the sources and deltas used when loading an asset are determined at load time, so that modifications to deltas/overrides/supplements are processed correctly.