Switched to PhantomReference based method for disposal of garbage collected assets.
This avoids the need for multiple garbage collection passes to handle asset classes.
With this change the Asset::onDispose() method is removed. Assets are now expected to register an static Runnable implementation with a provided DisposalHook, which will be run either when an asset is explicitly disposed, or if it is disposed after garbage collection via AssetType::processDisposal()